requirements.txt
flask
pywebview
openai
pdf2image
pytesseract
python-pptx
pillow
pydantic
